Refactor Poincaré event imports to align with core structure
- Updated import paths for `_PlaneEvent` and `_SurfaceEvent` to reflect their new location in the core events module.
- Removed the deprecated `events.py` file, streamlining the codebase and enhancing maintainability.
- Ensured consistency across various modules by adjusting import statements accordingly.
