+++ src/hiten/algorithms/poincare/core/events.py
