+++ src/hiten/algorithms/poincare/core/seeding.py
@@ -3,7 +3,7 @@ from typing import Protocol, runtime_checkable
 import numpy as np
 
 from hiten.algorithms.dynamics.base import _DynamicalSystemProtocol
-from hiten.algorithms.poincare.events import _SurfaceEvent
+from hiten.algorithms.poincare.core.events import _SurfaceEvent
 
 
 @runtime_checkable
