+++ src/hiten/algorithms/poincare/sh/backend.py
@@ -6,7 +6,7 @@ from scipy.optimize import root_scalar
 from hiten.algorithms.dynamics.base import (_DynamicalSystemProtocol,
                                             _propagate_dynsys)
 from hiten.algorithms.poincare.core.backend import _ReturnMapBackend
-from hiten.algorithms.poincare.events import (_PlaneEvent, _SectionHit,
+from hiten.algorithms.poincare.core.events import (_PlaneEvent, _SectionHit,
                                               _SurfaceEvent)
